Refactor server rules API helpers
